refactor: move filter modal into search form
