+++ search/src/components/SearchForm.tsx
@@ -5,6 +5,7 @@ import {
   Show,
   Switch,
   createEffect,
+  createMemo,
   createSignal,
   useContext,
 } from "solid-js";
@@ -17,9 +18,15 @@ import {
 } from "solid-headless";
 import { FaSolidCheck } from "solid-icons/fa";
 import { DatasetAndUserContext } from "./Contexts/DatasetAndUserContext";
-import { FilterModal } from "./FilterModal";
+import { Filter, FilterItem, Filters } from "./FilterModal";
 import { FiChevronDown, FiChevronUp } from "solid-icons/fi";
 import { SearchOptions, SearchStore } from "../hooks/useSearch";
+import { Tooltip } from "./Atoms/Tooltip";
+import { BsQuestionCircle } from "solid-icons/bs";
+
+const defaultFilter = {
+  field: "",
+};
 
 const SearchForm = (props: { search: SearchStore; groupID?: string }) => {
   const datasetAndUserContext = useContext(DatasetAndUserContext);
@@ -30,6 +37,44 @@ const SearchForm = (props: { search: SearchStore; groupID?: string }) => {
     // eslint-disable-next-line solid/reactivity
     props.search.state,
   );
+  const [tempFilterType, setTempFilterType] = createSignal<string>("must");
+  const [mustFilters, setMustFilters] = createSignal<Filter[]>([]);
+  const [mustNotFilters, setMustNotFilters] = createSignal<Filter[]>([]);
+  const [shouldFilters, setShouldFilters] = createSignal<Filter[]>([]);
+  const [jsonbPrefilter, setJsonbPrefilter] = createSignal<boolean>(true);
+  const curDatasetFiltersKey = createMemo(
+    () =>
+      `filters-${datasetAndUserContext.currentDataset?.()?.dataset.id ?? ""}`,
+  );
+
+  const saveFilters = (setShowFilterModal: (filter: boolean) => void) => {
+    const filters = {
+      must: mustFilters(),
+      must_not: mustNotFilters(),
+      should: shouldFilters(),
+      jsonb_prefilter: jsonbPrefilter(),
+    };
+    localStorage.setItem(curDatasetFiltersKey(), JSON.stringify(filters));
+    window.dispatchEvent(new Event("filtersUpdated"));
+    setShowFilterModal(false);
+  };
+
+  createEffect((prevFiltersKey) => {
+    const filtersKey = curDatasetFiltersKey();
+    if (prevFiltersKey === filtersKey) {
+      return filtersKey;
+    }
+
+    const savedFilters = localStorage.getItem(filtersKey);
+    if (savedFilters) {
+      const parsedFilters = JSON.parse(savedFilters) as Filters;
+      setMustFilters(parsedFilters.must);
+      setMustNotFilters(parsedFilters.must_not);
+      setShouldFilters(parsedFilters.should);
+      setJsonbPrefilter(parsedFilters.jsonb_prefilter ?? true);
+    }
+  }, "");
+
   const [searchTypes, setSearchTypes] = createSignal([
     { name: "Hybrid", isSelected: false, route: "hybrid" },
     {
@@ -127,7 +172,24 @@ const SearchForm = (props: { search: SearchStore; groupID?: string }) => {
             </div>
           </div>
           <div class="flex space-x-3">
-            <Popover defaultOpen={false} class="relative">
+            <Popover
+              defaultOpen={false}
+              class="relative"
+              onOpen={() => {
+                const filtersKey = curDatasetFiltersKey();
+                const savedFilters = localStorage.getItem(filtersKey);
+                if (savedFilters) {
+                  const parsedFilters = JSON.parse(savedFilters) as Filters;
+                  setMustFilters(parsedFilters.must);
+                  setMustNotFilters(parsedFilters.must_not);
+                  setShouldFilters(parsedFilters.should);
+                  setJsonbPrefilter(parsedFilters.jsonb_prefilter ?? true);
+                }
+              }}
+              onClose={() => {
+                saveFilters(() => {});
+              }}
+            >
               {({ isOpen, setState }) => (
                 <>
                   <PopoverButton
@@ -151,10 +213,202 @@ const SearchForm = (props: { search: SearchStore; groupID?: string }) => {
                       unmount={false}
                       class="absolute z-10 mt-2 h-fit w-fit rounded-md bg-neutral-200 p-1 shadow-lg dark:bg-neutral-700"
                     >
-                      <FilterModal
-                        showFilterModal={isOpen}
-                        setShowFilterModal={setState}
-                      />
+                      <div class="flex max-h-[50vh] min-w-[70vw] max-w-[75vw] flex-col space-y-2 overflow-auto px-2 pr-2 scrollbar-thin scrollbar-track-neutral-200 scrollbar-thumb-neutral-400 scrollbar-thumb-rounded-md dark:text-white dark:scrollbar-track-neutral-800 dark:scrollbar-thumb-neutral-600 xl:min-w-[50vw] 2xl:min-w-[40vw]">
+                        <div class="flex w-full items-center space-x-2 border-b border-neutral-400 py-2 dark:border-neutral-900">
+                          <label aria-label="Change Filter Type">
+                            <span class="p-1">Filter Type:</span>
+                          </label>
+                          <select
+                            class="h-fit rounded-md border border-neutral-400 bg-neutral-100 p-1 dark:border-neutral-900 dark:bg-neutral-800"
+                            onChange={(e) => {
+                              setTempFilterType(e.currentTarget.value);
+                            }}
+                            value={tempFilterType()}
+                          >
+                            <For each={["must", "must not", "should"]}>
+                              {(filter_type) => {
+                                return (
+                                  <option
+                                    classList={{
+                                      "flex w-full items-center justify-between rounded p-1":
+                                        true,
+                                      "bg-neutral-300 dark:bg-neutral-900":
+                                        filter_type === tempFilterType(),
+                                    }}
+                                  >
+                                    {filter_type}
+                                  </option>
+                                );
+                              }}
+                            </For>
+                          </select>
+                          <button
+                            type="button"
+                            class="rounded-md border border-neutral-400 bg-neutral-100 p-1 dark:border-neutral-900 dark:bg-neutral-800"
+                            onClick={() => {
+                              const curFilterType = tempFilterType();
+                              if (curFilterType === "must") {
+                                setMustFilters([
+                                  ...mustFilters(),
+                                  defaultFilter,
+                                ]);
+                              }
+                              if (curFilterType === "must not") {
+                                setMustNotFilters([
+                                  ...mustNotFilters(),
+                                  defaultFilter,
+                                ]);
+                              }
+                              if (curFilterType === "should") {
+                                setShouldFilters([
+                                  ...shouldFilters(),
+                                  defaultFilter,
+                                ]);
+                              }
+                            }}
+                          >
+                            + Add Filter
+                          </button>
+                          <div class="flex-1" />
+                          <label
+                            aria-label="Change JSONB Prefilter"
+                            class="flex items-center gap-x-1"
+                          >
+                            <Tooltip
+                              body={
+                                <BsQuestionCircle class="h-4 w-4 rounded-full fill-current" />
+                              }
+                              tooltipText="Only uncheck if on the enterprise plan and you wish to use custom indices for metadata filters."
+                            />
+                            <span>JSONB Prefilter:</span>
+                          </label>
+                          <input
+                            type="checkbox"
+                            class="rounded-md border border-neutral-400 bg-neutral-100 dark:border-neutral-900 dark:bg-neutral-800"
+                            onChange={(e) => {
+                              setJsonbPrefilter(e.currentTarget.checked);
+                            }}
+                            checked={jsonbPrefilter()}
+                          />
+                          <button
+                            type="button"
+                            class="rounded-md border border-neutral-400 bg-neutral-100 p-1 dark:border-neutral-900 dark:bg-neutral-800"
+                            onClick={() => {
+                              setMustFilters([]);
+                              setMustNotFilters([]);
+                              setShouldFilters([]);
+                            }}
+                          >
+                            Reset Filters
+                          </button>
+                          <button
+                            type="button"
+                            class="rounded-md border border-neutral-400 bg-neutral-100 p-1 dark:border-neutral-900 dark:bg-neutral-800"
+                            onClick={() => saveFilters(setState)}
+                          >
+                            Apply Filters
+                          </button>
+                        </div>
+                        <Show when={mustFilters().length > 0}>
+                          <div class="border-b border-neutral-400 py-2 dark:border-neutral-900">
+                            must: [
+                            <div class="flex flex-col gap-y-2">
+                              <For each={mustFilters()}>
+                                {(filter, index) => {
+                                  const onFilterChange = (
+                                    newFilter: Filter,
+                                  ) => {
+                                    const newFilters = mustFilters();
+                                    newFilters[index()] = newFilter;
+                                    setMustFilters(newFilters);
+                                  };
+
+                                  return (
+                                    <div
+                                      classList={{
+                                        "border-b border-dotted border-neutral-400 dark:border-neutral-900":
+                                          index() < mustFilters().length - 1,
+                                      }}
+                                    >
+                                      <FilterItem
+                                        initialFilter={filter}
+                                        onFilterChange={onFilterChange}
+                                      />
+                                    </div>
+                                  );
+                                }}
+                              </For>
+                            </div>
+                            ]
+                          </div>
+                        </Show>
+                        <Show when={mustNotFilters().length > 0}>
+                          <div class="border-b border-neutral-400 py-2 dark:border-neutral-900">
+                            must not: [
+                            <div class="flex flex-col gap-y-2">
+                              <For each={mustNotFilters()}>
+                                {(filter, index) => {
+                                  const onFilterChange = (
+                                    newFilter: Filter,
+                                  ) => {
+                                    const newFilters = mustNotFilters();
+                                    newFilters[index()] = newFilter;
+                                    setMustNotFilters(newFilters);
+                                  };
+
+                                  return (
+                                    <div
+                                      classList={{
+                                        "border-b border-dotted border-neutral-400 dark:border-neutral-900":
+                                          index() < mustNotFilters().length - 1,
+                                      }}
+                                    >
+                                      <FilterItem
+                                        initialFilter={filter}
+                                        onFilterChange={onFilterChange}
+                                      />
+                                    </div>
+                                  );
+                                }}
+                              </For>
+                            </div>
+                            ]
+                          </div>
+                        </Show>
+                        <Show when={shouldFilters().length > 0}>
+                          <div class="border-b border-neutral-400 py-2 dark:border-neutral-900">
+                            should: [
+                            <div class="flex flex-col gap-y-2">
+                              <For each={shouldFilters()}>
+                                {(filter, index) => {
+                                  const onFilterChange = (
+                                    newFilter: Filter,
+                                  ) => {
+                                    const newFilters = shouldFilters();
+                                    newFilters[index()] = newFilter;
+                                    setShouldFilters(newFilters);
+                                  };
+
+                                  return (
+                                    <div
+                                      classList={{
+                                        "border-b border-dotted border-neutral-400 dark:border-neutral-900":
+                                          index() < shouldFilters().length - 1,
+                                      }}
+                                    >
+                                      <FilterItem
+                                        initialFilter={filter}
+                                        onFilterChange={onFilterChange}
+                                      />
+                                    </div>
+                                  );
+                                }}
+                              </For>
+                            </div>
+                            ]
+                          </div>
+                        </Show>
+                      </div>
                     </PopoverPanel>
                   </Show>
                 </>
